+++ common/src/main/java/earth/terrarium/adastra/common/blockentities/base/RecipeMachineBlockEntity.java
@@ -25,9 +25,13 @@ public abstract class RecipeMachineBlockEntity<T extends Recipe<?>> extends Ener
         if (recipe != null && canFunction()) {
             recipeTick(level, getEnergyStorage());
         }
-        if (time % 5 == 0) {
+        if (time % 5 == 0 && shouldAutomaticallyUpdateLitState()) {
             setLit(cookTimeTotal > 0 && recipe != null && canFunction());
         }
+    }
+
+    public boolean shouldAutomaticallyUpdateLitState() {
+        return true;
     }
 
     @Override