+++ forge/src/main/java/earth/terrarium/adastra/datagen/provider/client/ModItemModelProvider.java
@@ -5,6 +5,7 @@ import com.teamresourceful.resourcefullib.common.registry.RegistryEntry;
 import earth.terrarium.adastra.AdAstra;
 import earth.terrarium.adastra.common.registry.ModItems;
 import net.minecraft.data.PackOutput;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.Item;
 import net.minecraftforge.client.model.generators.ItemModelProvider;
 import net.minecraftforge.client.model.generators.ModelFile;
@@ -14,6 +15,8 @@ import net.minecraftforge.registries.ForgeRegistries;
 import java.util.Objects;
 
 public class ModItemModelProvider extends ItemModelProvider {
+    public static final ResourceLocation RENDERED_ITEM = new ResourceLocation(AdAstra.MOD_ID, "item/rendered_item");
+
     public ModItemModelProvider(PackOutput output, ExistingFileHelper existingFileHelper) {
         super(output, AdAstra.MOD_ID, existingFileHelper);
     }