+++ forge/src/main/java/earth/terrarium/adastra/mixins/forge/common/RenderedItemMixin.java
@@ -3,6 +3,7 @@ package earth.terrarium.adastra.mixins.forge.common;
 import earth.terrarium.adastra.client.forge.AdAstraClientForge;
 import earth.terrarium.adastra.common.items.rendered.RenderedBlockItem;
 import earth.terrarium.adastra.common.items.rendered.RenderedItem;
+import earth.terrarium.adastra.common.items.rendered.TooltipRenderedBlockItem;
 import net.minecraft.client.renderer.BlockEntityWithoutLevelRenderer;
 import net.minecraft.world.item.Item;
 import net.minecraftforge.client.extensions.common.IClientItemExtensions;
@@ -10,7 +11,7 @@ import org.spongepowered.asm.mixin.Mixin;
 
 import java.util.function.Consumer;
 
-@Mixin({RenderedItem.class, RenderedBlockItem.class})
+@Mixin({RenderedItem.class, RenderedBlockItem.class, TooltipRenderedBlockItem.class})
 public abstract class RenderedItemMixin extends Item {
 
     public RenderedItemMixin(Properties properties) {